Apply changes from my fork
