+++ src/Viberaria/Config/VibrationPattern.cs
