+++ src/Viberaria/IntifaceConnection/Vibration.cs
