+++ src/Viberaria/tModAdapters/tChat.cs
@@ -1,30 +1,39 @@
 using System;
+using log4net;
 using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.Chat;
 using Terraria.Localization;
-using static Viberaria.ViberariaConfig;
+using Terraria.ModLoader;
+using static Viberaria.Config.ViberariaConfig;
 
-namespace Viberaria;
+namespace Viberaria.tModAdapters;
 
 public static class tChat
 {
+    public static ILog Logger => ModContent.GetInstance<tMod>().Logger;
+
     private static NetworkText LiteralText(string msg)
     {
         return NetworkText.FromLiteral(msg);
     }
 
+    /// <summary>
+    /// Send a message to the player client in chat with a given color.
+    /// </summary>
+    /// <param name="msg">The message to send. Is automatically prefixed with the timestamp when the Debug config has been enabled.</param>
+    /// <param name="color">The color the message should have.</param>
     public static void LogToPlayer(string msg, Color color)
     {
-        if (!Main.dedServ)
+        if (Main.dedServ) return;
+
+        if (Instance.Debug.Enabled)
         {
-            if (Instance.DebugChatMessages)
-            {
-                string time = DateTime.Now.Minute + ":" +
-                              (DateTime.Now.Second + DateTime.Now.Millisecond * 0.001f).ToString("0.00");
-                msg = time + "| " + msg;
-            }
-            ChatHelper.SendChatMessageToClient(LiteralText(msg), color, Main.myPlayer);
+            string time = DateTime.Now.Minute + ":" +
+                          (DateTime.Now.Second + DateTime.Now.Millisecond * 0.001f).ToString("00.00");
+            msg = time + "| " + msg;
         }
+
+        ChatHelper.SendChatMessageToClient(LiteralText(msg), color, Main.myPlayer);
     }
 }
